Crate chalk_derive

Enums

DeriveKind

Functions

bounded_by_trait

derive_any_visit

derive_fold
Derives Fold for structs and enums for which one of the following is true:

derive_has_interner

derive_super_visit
Same as Visit, but derives SuperVisit instead

derive_visit
Derives Visit for structs and enums for which one of the following is true:

derive_zip

each_variant_pair

find_interner

get_generic_param

get_generic_param_name

has_interner
Checks whether a generic parameter has a : HasInterner bound

has_interner_attr

is_interner
Checks whether a generic parameter has a : Interner bound

Derive Macros

Fold

HasInterner

SuperVisit

Visit

Zip