Crate chalk_derive

Enums

DeriveKind ğŸ”’

Functions

bounded_by_trait ğŸ”’

derive_any_type_visitable ğŸ”’

derive_fallible_type_folder ğŸ”’

derive_has_interner ğŸ”’

derive_type_foldable ğŸ”’

Derives TypeFoldable for structs and enums for which one of the following is true:

derive_type_super_visitable ğŸ”’

Same as TypeVisitable, but derives TypeSuperVisitable instead

derive_type_visitable ğŸ”’

Derives TypeVisitable for structs and enums for which one of the following is true:

derive_zip ğŸ”’

each_variant_pair ğŸ”’

find_interner ğŸ”’

get_intern_param ğŸ”’

get_intern_param_name ğŸ”’

has_interner ğŸ”’

Checks whether a generic parameter has a : HasInterner bound

has_interner_attr ğŸ”’

is_interner ğŸ”’

Checks whether a generic parameter has a : Interner bound

try_find_interner ğŸ”’

Derive Macros

FallibleTypeFolder

HasInterner

TypeFoldable

TypeSuperVisitable

TypeVisitable

Zip